--- src/patterns/pages/layout/components/aside.component.ts
@@ -1,14 +1,0 @@
-// import {
-//     Component,
-//     ContentChild,
-//     AfterContentInit,
-//     HostBinding
-// } from '@angular/core';
-
-// @Component({
-//     selector: 'sam-aside',
-//     template: '<ng-content></ng-content>'
-// })
-// export class SamAsideComponent {
-//     @HostBinding('class.sidebar') container: boolean = true;
-// }
