+++ src/patterns/pages/layout/components/aside/index.ts
@@ -1,0 +1,3 @@
+export * from './aside-container.component';
+export * from './aside.component';
+export * from './aside.module';
