+++ src/patterns/pages/layout/components/index.ts
@@ -1,5 +1,7 @@
 export * from './actionbar.component';
+export * from './aside';
 export * from './layout.component';
 export * from './main.component';
 export * from './pagination.component';
 export * from './toolbar.component';
+export * from './layout-components.module';
