+++ src/patterns/pages/layout/components/layout.component.ts
@@ -15,96 +15,7 @@ import {
 } from '@angular/core';
 import { SamActionBarComponent } from './actionbar.component';
 import { SamMainComponent } from './main.component';
-
-@Component({
-  selector: 'sam-aside',
-  template: '<ng-content></ng-content>'
-})
-export class SamAsideComponent {
-  @HostBinding('class') public classes = 'sam-aside';
-  @Input()
-  public get opened () {
-    return this._opened;
-  }
-
-  public set opened (v: boolean) {
-    this.toggle(!!v);
-  }
-
-  @Output() public onOpen = new EventEmitter<void>();
-  @Output() public onClose = new EventEmitter<void>();
-
-  private _opened: boolean;
-
-  public open (): void {
-    return this.toggle(true);
-  }
-
-  public close (): void {
-    return this.toggle(false);
-  }
-
-  public toggle (isOpen: boolean = !this.opened): void {
-    if (isOpen === this._opened) {
-      return;
-    } else {
-      this._opened = isOpen;
-
-      if (this._opened === true) {
-        this.onOpen.emit();
-      } else {
-        this.onClose.emit();
-      }
-    }
-  }
-}
-
-@Component({
-  selector: 'sam-aside-container',
-  template: '<ng-content></ng-content>'
-})
-export class SamAsideContainerComponent {
-
-  @HostBinding('class')
-  public classes = 'sidebar sam-aside-container';
-
-  @ContentChildren(forwardRef(() => SamAsideComponent), { descendants: true})
-  public sidenavs;
-
-  constructor (
-    private _element: ElementRef,
-    private _renderer: Renderer2) {}
-
-  public registerSidenav (sidenav: SamAsideComponent) {
-    if (!sidenav) {
-      return;
-    } else {
-
-      sidenav.onOpen.subscribe(
-        () => this._setContainerClass(true)
-      );
-
-      sidenav.onClose.subscribe(
-        () => this._setContainerClass(false)
-      );
-    }
-  }
-
-  private _setContainerClass (isAdd: boolean) {
-    if (isAdd) {
-      this._renderer.addClass(
-        this._element.nativeElement,
-        'aside-opened'
-      );
-    } else {
-      this._renderer.removeClass(
-        this._element.nativeElement,
-        'aside-opened'
-      );
-    }
-  }
-
-}
+import { SamAsideContainerComponent, SamAsideComponent } from './aside';
 
 @Component({
   selector: 'sam-layout-b',
