+++ src/patterns/pages/layout/index.ts
@@ -1,7 +1,8 @@
 export * from './layout.component';
-export * from './components/toolbar.component';
-export * from './components/actionbar.component';
-// export * from './components/aside.component';
-export * from './components/main.component';
-export * from './components/layout.component';
-export * from './components/pagination.component';
+export * from './components';
+// export * from './components/toolbar.component';
+// export * from './components/actionbar.component';
+// export * from './components/main.component';
+// export * from './components/layout.component';
+// export * from './components/pagination.component';
+// export * from './components/aside';
